* separated tests into smaller files * created fixture for test data
